requirements.txt
google-api-core==2.18.0
google-auth==2.29.0
google-cloud-aiplatform==1.46.0
google-cloud-bigquery==3.17.2
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.3
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
langchain==0.1.14
langchain-cli==0.0.21
langchain-community==0.0.31
langchain-core==0.1.40
langchain-experimental==0.0.56
langchain-google-vertexai==0.1.2
streamlit==1.31.1
db-dtypes